- MIDI ticks-per-beat snap resolution fix. (bis)

+++ src/qtractorAudioEngine.h
@@ -254,7 +254,7 @@ public:
 		bool           playing;
 		unsigned int   sampleRate;
 		float          tempo;
-		unsigned short ticksPerBeat;
+		unsigned int   ticksPerBeat;
 		unsigned short beatsPerBar;
 		unsigned short beatType;
 		float          beats;
