+++ src/qtractorTimeScale.h
@@ -227,7 +227,7 @@ public:
 
 		// Alternate (secondary) time-sig helper methods
 		unsigned short beatsPerBar2() const;
-		unsigned short ticksPerBeat2() const;
+		unsigned int ticksPerBeat2() const;
 
 		// Node keys.
 		unsigned long  frame;
@@ -242,7 +242,7 @@ public:
 		unsigned short beatsPerBar;
 		unsigned short beatDivisor;
 
-		unsigned short ticksPerBeat;
+		unsigned int   ticksPerBeat;
 
 	protected:
